pyproject.toml
[project]
name = "mcp-demo-starlette"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "load-dotenv>=0.1.0",
    "mcp[cli]>=1.7.1",
    "pydantic-ai>=0.1.10",
    "starlette>=0.46.2",
]
